requirements.txt
# Google API Dependencies
google-api-python-client>=2.0.0  # Core Google API client library
google-auth-httplib2>=0.1.0      # HTTP client for authentication
google-auth-oauthlib>=1.0.0      # OAuth 2.0 library

# AI Dependencies
anthropic>=0.3.0

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# Configuration
python-dotenv>=0.19.0
pyyaml>=6.0.0

# Utilities
rich>=13.0.0  # For better logging output
tenacity>=8.0.0  # For retry logic

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0